backend-service/pyproject.toml
[project]
name = "backend-service"
version = "0.1.0"
description = "Backend Service for the OpenTelemetry tracing demo"
requires-python = ">=3.12"
dependencies = [
    "flask",
    "opentelemetry-distro",
    "opentelemetry-exporter-otlp",
    "opentelemetry-instrumentation-flask",
    "opentelemetry-instrumentation-requests",
]
